refactor(codebase): 👷 naming and folders
