+++ components/pages/home/hero/HeroMainLinks.tsx
@@ -1,8 +1,9 @@
-import HeroLink from '@/components/homepage/hero/HeroLink'
+import HeroLink from '@/components/pages/home/hero/HeroLink'
+
+import { TEXT } from '@/localization/english'
 
 import { DATA_TEST_IDS } from '@/__tests__/playwright/lib/utils/constants/ids/dataTestIds'
 import { EXTERNAL_URL } from '@/lib/utils/constants/urls/externalUrls'
-import { TEXT } from '@/localization/english'
 
 const HeroMainLinks = (): JSX.Element => {
   return (
