+++ components/pages/home/skills/SkillsForCompany.tsx
