+++ components/pages/home/skills/SkillsMain.tsx
