+++ components/pages/resume/ResumeInfo.tsx
@@ -2,6 +2,7 @@ import Heading1 from '@/components/shared/Heading1'
 import Highlight from '@/components/shared/Highlight'
 
 import { DATA_TEST_IDS } from '@/__tests__/playwright/lib/utils/constants/ids/dataTestIds'
+
 import { TEXT } from '@/localization/english'
 
 const ResumeInfo = (): JSX.Element => {
