+++ lib/data/expertise/mindset.ts
@@ -6,20 +6,20 @@ export const mindset: SkillInfo[] = [
     order: 1,
     title: 'Continuous learning',
     description:
-      'I continuously improve my technical skills and stay updated with trends by actively engaging in online communities and\u00A0learning to\u00A0enhance my knowledge and\u00A0expertise.',
+      'I continuously improve my technical skills and stay updated with trends by actively learning to\u00A0enhance my knowledge and\u00A0expertise.',
   },
   {
     id: 1,
     order: 2,
     title: 'Team player',
     description:
-      'I enjoy working in teams where I can contribute ideas. My strong communication skills enable me to explain concepts effectively to\u00A0stakeholders and\u00A0team members.',
+      'I like teams where I can contribute ideas. My communication skills enable me to explain concepts to\u00A0stakeholders and\u00A0team members.',
   },
   {
     id: 2,
     order: 3,
     title: 'Delivering meaningful products',
     description:
-      'I am passionate about solving real-world problems to help users and customers. Products that streamline processes, and\u00A0increase value of a\u00A0business.',
+      'I enjoy solving real-world problems to help users and customers. I\u00A0like to streamline processes, and\u00A0increase value of a\u00A0business.',
   },
 ]
